Crate terbium_grammar

Re-exports

pub use crate::ast::Body;

pub use crate::ast::Expr;

pub use crate::ast::Node;

pub use crate::ast::Param;

pub use crate::ast::ParseInterface;

pub use crate::ast::Target;

pub use crate::ast::TypeExpr;

pub use crate::error::*;

pub use crate::token::get_lexer as tokenizer;

pub use crate::token::Operator;

pub use crate::token::Token;

Modules

ast

error

token

Structs

ChumskyStream

A type that represents a stream of input tokens. Unlike Iterator, this type supports backtracking and a few other features required by the crate.

Source

Span

Spanned

Traits

ChumskyParser

A trait implemented by parsers.